refactor(mitm-addon): keep websocket helpers narrow
