+++ crates/runner/mitm-addon/tests/test_model_provider_websocket_usage.py
@@ -2,6 +2,8 @@
 
 import json
 import uuid
+from collections.abc import Callable
+from pathlib import Path
 
 import pytest
 from mitmproxy import http
@@ -16,14 +18,13 @@ from tests.model_provider_websocket_helpers import (
     _feed_websocket_server_message,
     _model_websocket_usage_sources,
     _openai_model_websocket_flow,
-    _openai_model_websocket_request_flow,
     _openai_websocket_usage_frame,
     _run_deferred_websocket_trims,
     _ScheduledWebSocketTrim,
     _set_websocket_message,
-    _write_openai_model_websocket_registry,
 )
 from tests.pending_helpers import assert_pending
+from tests.request_handler_helpers import _single_firewall_vm, _write_registry
 
 
 @pytest.fixture(autouse=True)
@@ -31,6 +32,44 @@ def deferred_websocket_trim_scheduler(
     monkeypatch: pytest.MonkeyPatch,
 ) -> list[_ScheduledWebSocketTrim]:
     return _capture_deferred_websocket_trims(monkeypatch)
+
+
+def _write_openai_model_websocket_registry(tmp_path: Path) -> Path:
+    firewall_name = "model-provider:openai-api-key"
+    return _write_registry(
+        tmp_path,
+        vm_info=_single_firewall_vm(
+            tmp_path,
+            run_id="run-abc-123",
+            sandbox_marker="tok-xyz",
+            firewall_name=firewall_name,
+            api_entry={
+                "base": "https://api.openai.com",
+                "auth": {"headers": {"Authorization": "Bearer token"}},
+                "permissions": [{"name": "responses", "rules": ["POST /v1/responses"]}],
+            },
+            network_policy={
+                "allow": ["responses"],
+                "deny": [],
+                "ask": [],
+                "unknownPolicy": "deny",
+            },
+            billable_firewalls=[firewall_name],
+            vm_fields={"cliAgentType": "codex"},
+        ),
+    )
+
+
+def _openai_model_websocket_request_flow(
+    real_flow: Callable[..., http.HTTPFlow],
+) -> http.HTTPFlow:
+    return real_flow(
+        with_response=False,
+        client_ip="10.200.0.5",
+        host="api.openai.com",
+        path="/v1/responses",
+        method="POST",
+    )
 
 
 class TestModelProviderWebSocketUsage:
